ssh/tailssh: fix privilege dropping on FreeBSD
On FreeBSD and Darwin, changing a process's supplementary groups with
setgroups(2) will also change the egid of the process, setting it to the
first entry in the provided list. This is distinct from the behaviour on
other platforms (and possibly a violation of the POSIX standard).

Because of this, our incubator code wouldn't change the process's gid,
because it would read the newly-changed egid, compare it against the
expected egid, and since they matched, not change the gid.

This could be observed by running "id -p" in two contexts. The expected
output, and the output returned when running from a SSH shell, is:

    andrew@freebsd:~ $ id -p
    uid         andrew
    groups      andrew

However, when run via "ssh andrew@freebsd id -p", the output would be:

    $ ssh andrew@freebsd id -p
    login       root
    uid         andrew
    rgid        wheel
    groups      andrew

(this could also be observed via "id -g -r" to print just the gid)

We fix this by ensuring that the setGroups call behaves the same on
FreeBSD as it does on other platforms by adding the current egid as the
first entry in the list, and thus not changing it unexpectedly.

More information can be found in the following article:
    https://www.usenix.org/system/files/login/articles/325-tsafrir.pdf

Also, while we're here... defensively call runtime.LockOSThread to make
it less likely we trip over bugs like golang/go#1435

Signed-off-by: Andrew Dunham <andrew@du.nham.ca>
Change-Id: Icc08c629ddc87923aa9ae0d542f63925e9c40cd3