go.mod,cmd/mkpkg,packages,release/dist: update to goreleaser v2
The main change affecting us appears to be how files are specified in
the manifests. There are insufficient docs in goreleaser, so it's not
exactly clear if files.TypeDir in contents has the exact same behavior
as EmptyDirs, but the tests are now happy. We should manually test some
package builds ASAP.

Updates #8043

Signed-off-by: James Tucker <james@tailscale.com>