tstest: add a helper for reproducable random seeds
We recently ran across a case of flaky tests that were seeding the
math/rand package and using that to generate data. The tests were
flaking on specific random values that were difficult to reproduce. If
the test had been using a helper like this, the flaky cases could have
been reliably reproduced.

Signed-off-by: James Tucker <james@tailscale.com>