portlist: refactor Poller to simplify external API
This change makes it so that Poller has only two external methods: Run and Close.
The Poller becomes harder to get wrong by external packages because Close is a no-op
if Run is not called and Run must be called to obtain updates and check an init-error.

The only 3rd method is an exported "Init" that preserves an old behavior whereby you can
pre-fetch a list of ports and check for errors _without_ starting Run's ticker.
However, I'd be happy to remove that if we can afford it.

Fixes #8171

Signed-off-by: Marwan Sulaiman <marwan@tailscale.com>