prober: make DERP probes fail when stun requests fail
The client treats a single lost stun request as a failure in netcheck
and will bounce regions as a result. The prober results should represent
this if it is happening regularly in a region we need to know about
that. The fist step is to report the data with more precision; the
likely immediate next step is to adjust alerting to soften the blow.

Updates tailscale/corp#11492

Signed-off-by: James Tucker <james@tailscale.com>