wgengine/netlog: fix nil pointer dereference in logtail
When `Logger.Startup` is called and a new `http.Client` is initialized
for logtail, `nl.logger` is nil. It's initialized right after creating
the `http.Client`. But because we're passing a method value, it actually
works and passes the value with a `nil` receiver.
Later, when `logtail.Logger.Logf` is called from the `http.Client`,
`tailscaled` panics when trying to access the fields of that nil logger.

Signed-off-by: Andrew Lytvynov <awly@tailscale.com>