version/distro: fallback to /etc/os-release on Linux
On most Linux distros, you can read the distro ID from
`/etc/os-release`. In fact, we do that in
https://github.com/tailscale/tailscale/blob/main/scripts/installer.sh#L38

Do this as a fallback to avoid accidentally breaking some logic
elsewhere. This will only change the return value where it was
previously an empty string.

Also add a few known distros to the const list.

Updates #6995

Signed-off-by: Andrew Lytvynov <awly@tailscale.com>