health, ipn/ipnlocal: add healthcheck for state/config
If the user sufficiently breaks their state file such that the local
machine or node keys are zero, the machine ends up in a broken state but
we don't surface this to the user. Add a healthcheck that tracks errors
when these key(s) are zero and surfaces them in 'tailscale status'/etc.

Signed-off-by: Andrew Dunham <andrew@du.nham.ca>
Change-Id: Ideffbf9510b09850ebf8daf440fac78109de7812