net/netns: indicate when android protect func is no longer nil
We have been getting into routing loops due to the timing of when we
bind sockets on starting the tailscale android app. At this point, we do
not have access to `VpnService.protect()` and are unable to protect
the magicsock sockets, which causes a routing loop issue until we
forcibly rebind about 10 minutes into the service being started.

This change returns a bool from SetAndroidProtectFunc we get access to
a func.

Updates tailscale/corp#13814

Signed-off-by: Charlotte Brandhorst-Satzkorn <charlotte@tailscale.com>