wgengine/magicsock: only accept pong sent by CLI ping
When sending a ping from the CLI, only accept a pong that is in reply
to this CLI ping we sent. Otherwise, CLI ping cannot be used to probe
MTU, since it will accept any heartbeat ping as a pong in reply.

XXX This version switches from DERP over to UDP-only once it gets any
direct connection, even if the UDP connection can't handle the size of
ping. Ideally it would instead continue to ping all paths until it got
a direct connection that can handle this size of ping.

Signed-off-by: Val <valerie@tailscale.com>